+++ src/solver/simulation/include/antares/solver/simulation/sim_structure_probleme_economique.h
@@ -75,6 +75,9 @@ struct CORRESPONDANCES_DES_VARIABLES
         std::vector<int> CostVariationWithdrawal;
         std::vector<int> OverflowVariable;
     } SIM_ShortTermStorage;
+
+    std::vector<int> BindingConstraintPenaltyPos;
+    std::vector<int> BindingConstraintPenaltyNeg;
 };
 
 struct CORRESPONDANCES_DES_CONTRAINTES