+++ src/solver/simulation/sim_alloc_probleme_hebdo.cpp
@@ -187,6 +187,8 @@ void SIM_AllocationProblemePasDeTemps(PROBLEME_HEBDO& problem,
                                                                             0);
         variablesMapping.SIM_ShortTermStorage.CostVariationWithdrawal.assign(shortTermStorageCount,
                                                                              0);
+        variablesMapping.BindingConstraintPenaltyPos.assign(activeConstraints.size(), 0);
+        variablesMapping.BindingConstraintPenaltyNeg.assign(activeConstraints.size(), 0);
 
         variablesMapping.SIM_ShortTermStorage.OverflowVariable.assign(shortTermStorageCount, 0);